Move frames outside persona/
