+++ _api/frames/MintSuccessFrame.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
 
-import { OGAvatar } from '../../og/OGAvatar';
-import { getViewerFromParams } from '../getViewerFromParams';
-import { FrameBgImage, IMAGE_URL_BASE } from '../layoutFragments/FrameBgImage';
-import { FrameBody } from '../layoutFragments/FrameBody';
-import { FrameBodyGradient } from '../layoutFragments/FrameBodyGradient';
-import { FrameFooter } from '../layoutFragments/FrameFooter';
-import { FrameHeadline } from '../layoutFragments/FrameHeadline';
-import { FrameWrapper } from '../layoutFragments/FrameWrapper';
-import { Frame } from '../router';
-import { staticResourceIdentifier } from '../staticResourceIdentifier';
+import { OGAvatar } from '../og/OGAvatar';
+
+import { getViewerFromParams } from './getViewerFromParams';
+import { FrameBgImage, IMAGE_URL_BASE } from './layoutFragments/FrameBgImage';
+import { FrameBody } from './layoutFragments/FrameBody';
+import { FrameBodyGradient } from './layoutFragments/FrameBodyGradient';
+import { FrameFooter } from './layoutFragments/FrameFooter';
+import { FrameHeadline } from './layoutFragments/FrameHeadline';
+import { FrameWrapper } from './layoutFragments/FrameWrapper';
+import { Frame } from './router';
+import { staticResourceIdentifier } from './staticResourceIdentifier';
 
 const imageNode = async (params: Record<string, string>) => {
   const { viewerProfile } = await getViewerFromParams(params);
